chore(deps) & fixed various bugs
