+++ next.config.mjs
@@ -13,24 +13,13 @@ const nextConfig = {
   swcMinify: true,
   output: process.env.BUILD_STANDALONE === "true" ? "standalone" : undefined,
   basePath: "/beta",
-  experimental: {
-    appDir: true,
-    extensionAlias: {
-      ".js": [".js", ".jsx", ".ts", ".tsx"],
-    },
-  },
   modularizeImports: {
-    "@components/index.(js|ts)": {
-      transform: "@components/{{member}}",
+    "@/components/?(((\\w*)?/?)*)": {
+      transform: "@/components/{{ matches.[1] }}/{{member}}",
       preventFullImport: true,
       skipDefaultConversion: true,
     },
-    "@schemas/index.(js|ts)": {
-      transform: "@schemas/{{member}}",
-      preventFullImport: true,
-      skipDefaultConversion: true,
-    },
-    "@utils/?(((\\w*)?/?)*)/index.(js|ts)": {
+    "@utils/?(((\\w*)?/?)*)": {
       transform: "@utils/{{ matches.[1] }}/{{member}}",
       preventFullImport: true,
       skipDefaultConversion: true,
