+++ package.json
@@ -2,6 +2,7 @@
   "name": "next-app",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "analyze": "cross-env ANALYZE=true pnpm build",
     "build": "next build",
@@ -13,47 +14,47 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@headlessui/react": "1.7.14",
+    "@headlessui/react": "1.7.15",
     "@hookform/resolvers": "3.1.0",
     "@monaco-editor/react": "^4.5.1",
-    "@next/bundle-analyzer": "13.4.2-canary.5",
+    "@next/bundle-analyzer": "13.4.4",
     "@reduxjs/toolkit": "1.9.5",
     "client-only": "0.0.1",
     "cross-env": "7.0.3",
-    "next": "13.4.4-canary.0",
+    "next": "13.4.4",
     "next-redux-wrapper": "8.1.0",
-    "react": "18.3.0-canary-7cd98ef2b-20230509",
-    "react-dom": "18.3.0-canary-7cd98ef2b-20230509",
-    "react-hook-form": "7.43.9",
-    "react-redux": "8.0.5",
+    "react": "18.3.0-canary-21a161fa3-20230609",
+    "react-dom": "18.3.0-canary-21a161fa3-20230609",
+    "react-hook-form": "7.44.3",
+    "react-redux": "8.0.7",
     "redux": "4.2.1",
     "server-only": "0.0.1",
     "zod": "3.21.4"
   },
   "devDependencies": {
-    "@types/node": "20.1.2",
-    "@types/react": "18.2.6",
+    "@types/node": "20.2.6",
+    "@types/react": "18.2.11",
     "@types/react-dom": "18.2.4",
     "@types/shell-quote": "1.7.1",
-    "@typescript-eslint/eslint-plugin": "5.59.5",
-    "@typescript-eslint/parser": "5.59.5",
+    "@typescript-eslint/eslint-plugin": "5.59.9",
+    "@typescript-eslint/parser": "5.59.9",
     "autoprefixer": "10.4.14",
-    "eslint": "8.40.0",
-    "eslint-config-next": "13.4.2-canary.5",
+    "eslint": "8.42.0",
+    "eslint-config-next": "13.4.4",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-simple-import-sort": "10.0.0",
-    "eslint-plugin-tailwindcss": "3.11.0",
+    "eslint-plugin-tailwindcss": "3.12.1",
     "eslint-plugin-tsdoc": "0.2.17",
     "husky": "8.0.3",
     "lint-staged": "13.2.2",
     "npm-check-updates": "16.10.12",
-    "postcss": "8.4.23",
+    "postcss": "8.4.24",
     "prettier": "2.8.8",
-    "prettier-plugin-tailwindcss": "0.2.8",
+    "prettier-plugin-tailwindcss": "0.3.0",
     "shell-quote": "1.8.1",
     "tailwindcss": "3.3.2",
-    "typescript": "5.1.0-dev.20230511"
+    "typescript": "5.1.3"
   },
   "packageManager": "pnpm@8.3.0"
 }
