+++ src/app/login/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Suspense } from "react";
 
-import { LoginForm } from "./LoginForm.js";
+import { LoginForm } from "./LoginForm";
 
 export const metadata: Metadata = {
   title: "Login",
