--- src/app/not-found.tsx
@@ -1,3 +1,0 @@
-const RootNotFound = () => <h1>Page not found</h1>;
-
-export default RootNotFound;
