+++ src/features/redux_thunk/hooks.ts
@@ -1,7 +1,7 @@
 import type { TypedUseSelectorHook } from "react-redux";
 import { useDispatch, useSelector } from "react-redux";
 
-import type { AppDispatch, RootState } from "./store.js";
+import type { AppDispatch, RootState } from "./store";
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch: () => AppDispatch = useDispatch;
