--- src/shared/schemas/index.ts
@@ -1,1 +1,0 @@
-export { loginSchema, type LoginSchemaType } from "./loginSchema.js";
