+++ src/shared/utils/shared/index.ts
@@ -1,2 +1,2 @@
-export type { ClassArray, ClassDictionary, ClassValue } from "./clsx.js";
-export { clsx } from "./clsx.js";
+export type { ClassArray, ClassDictionary, ClassValue } from "./clsx";
+export { clsx } from "./clsx";
